test(api): move the tls fixture back into the unix-only test module
